requirements.txt
Django>=5.2.5
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0




